[eagv2 session8] Working code with gmail integration
Increased max iterations to 10. Now the tool sends email to
specified gmail ID

Verified with following prompts:

[Prompt1]
Give me a list of Top 10 cricketers in the world based on ICC rankings for year 2025, send email to anurag2k@gmail.com with the results.

[Prompt2]
Give me a list of seven wonders of the world, send email to anurag2k@gmail.com with the results